+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.7.15]
 ### Fixed
 - Re-enable scan results when re-enabling different scan types
+- (LS Preview) do not trigger scan on startup for Snyk Code multiple times
 
 ## [2.7.14]
 ### Added